refactor: extract useDismissableMenu, shared by AccountMenu and ListToolbar
Both components hand-rolled the same outside-click/Escape listener
pair for their dropdowns. The shared hook also adds focus restoration
on Escape for AccountMenu's trigger, which the duplicated version
didn't have.
